duan

+++ ass2fontend/src/pages/Admin/User/AdminUser.tsx
@@ -1,6 +1,7 @@
-import React, { useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import { Users } from "../../../interface/User";
 import api from "../../../axios";
+import Swal from "sweetalert2";
 
 const AdminUser = () => {
   const [users, setUsers] = useState<Users[]>([]);
@@ -19,14 +20,26 @@ const AdminUser = () => {
   // Hàm xóa người dùng
   const deleteUser = async (id: string) => {
     try {
-      // Gửi yêu cầu xóa người dùng
-      await api.delete(`/user/${id}`);
-      // Cập nhật lại danh sách người dùng sau khi xóa
-      setUsers(users.filter((user) => user._id !== id));
-      alert("Xóa người dùng thành công!");
+      if (confirm("Are you sure you want to delete this user")) {
+        // Gửi yêu cầu xóa người dùng
+        await api.delete(`/user/${id}`);
+        // Cập nhật lại danh sách người dùng sau khi xóa
+        setUsers(users.filter((user) => user._id !== id));
+        Swal.fire({
+          title: "Success",
+          text: "Xóa người dùng thành công",
+          icon: "success",
+          confirmButtonText: "OK",
+        });
+      }
     } catch (error) {
-      console.error("Xóa người dùng thất bại:", error);
-      alert("Có lỗi xảy ra khi xóa người dùng!");
+      console.log(error);
+      Swal.fire({
+        title: "Error",
+        text: "Xóa thất bại",
+        icon: "error",
+        confirmButtonText: "OK",
+      });
     }
   };
 
@@ -36,6 +49,7 @@ const AdminUser = () => {
         <thead>
           <tr>
             <th>Username</th>
+            <th>Image</th>
             <th>Email</th>
             <th>Password</th>
             <th>Action</th>
@@ -45,12 +59,15 @@ const AdminUser = () => {
           {users.map((p) => (
             <tr key={p._id}>
               <td>{p.username}</td>
+              <td>
+                <img src={p.image} alt="" width={100} />
+              </td>
               <td>{p.email}</td>
               <td>{p.password}</td>
               <td>
                 <button
                   className="btn btn-danger"
-                  onClick={() => deleteUser(p._id)}
+                  onClick={() => deleteUser(p._id!)}
                 >
                   Xóa
                 </button>
